+++ Obelheiro/pesquisa/honeypot_data/rust/pcap_reader_pcap_parser/src/pcap/block/ntp.rs
@@ -3,8 +3,8 @@ use postgres::Transaction;
 // use rusqlite::{params, Connection};
 
 pub struct PcapNTP {
-  pub id: i64,
-  pub refid: i64,
+  pub id: i32,
+  pub refid: i32,
 }
 
 impl PcapNTP {
@@ -21,19 +21,19 @@ impl PcapNTP {
   }
 }
 
-pub fn process_ntp(ntp_packet: &NtpPacket, id: i64) -> PcapNTP {
+pub fn process_ntp(ntp_packet: &NtpPacket, id: i32) -> PcapNTP {
   match ntp_packet {
     ntp_parser::NtpPacket::V3(ntp_v3) => {
       return PcapNTP {
         id,
         // TODO: verify monlist check bytes
-        refid: ntp_v3.ref_id as i64,
+        refid: ntp_v3.ref_id as i32,
       };
     }
     ntp_parser::NtpPacket::V4(ntp_v4) => {
       return PcapNTP {
         id,
-        refid: ntp_v4.ref_id as i64,
+        refid: ntp_v4.ref_id as i32,
       };
     }
   }
